# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RMS Backend API
  description: >
    REST API untuk RMS (Ruas Management System) - manajemen purchase order,
    klien, dan tim lapangan. Hasil port dari monolith PHP ke Go + MySQL +
    Redis. Semua endpoint di bawah `/api/*` (kecuali `/api/auth/*` publik)
    butuh header `Authorization: Bearer <access_token>`. Endpoint mutating
    tertentu juga dibatasi RBAC lewat `access_slug` (lihat deskripsi
    masing-masing operasi).
  version: "1.1.0"
  contact:
    name: RMS Backend
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.rms.example.com
    description: Production (ganti sesuai domain Anda)
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login admin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login berhasil
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Email atau password salah / akun nonaktif
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
        "429":
          description: Terlalu banyak percobaan login (throttled)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotasi refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token baru diterbitkan
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TokenPairResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Refresh token tidak valid / sudah dipakai / expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
  /api/auth/activate:
    post:
      tags: [Auth]
      summary: Aktivasi akun admin baru dari email undangan
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateRequest"
      responses:
        "200":
          description: Akun berhasil diaktifkan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Minta link reset password
      description: >
        Selalu membalas sukses (terlepas email terdaftar atau tidak) untuk
        mencegah enumerasi email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Instruksi selalu dikirim jika email terdaftar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
  /api/auth/reset-password-code:
    post:
      tags: [Auth]
      summary: Reset password memakai kode dari email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetWithCodeRequest"
      responses:
        "200":
          description: Password berhasil direset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Identitas admin yang sedang login
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke sesi)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Logout berhasil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Ganti password sendiri (revoke semua sesi lain)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password berhasil diganti
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponseBase" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Password lama salah
        "403":
          description: Akun tidak aktif
        "429":
          description: Terlalu banyak percobaan gagal (throttled)

  # =========================================================
  # REGION
  # =========================================================
  /api/regions:
    get:
      tags: [Region]
      summary: List region (dengan paginasi)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/RegionListPaged"
    post:
      tags: [Region]
      summary: Buat region baru
      description: "Butuh access_slug: `create_region`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/regions/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathInt"
    get:
      tags: [Region]
      summary: Detail region
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/RegionResponse" }
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Region]
      summary: Ubah region
      description: "Butuh access_slug: `edit_region`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Region]
      summary: Hapus region
      description: "Butuh access_slug: `delete_region`"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          description: Gagal hapus (kemungkinan masih dipakai data lain)

  # =========================================================
  # DIVISION
  # =========================================================
  /api/divisions:
    get:
      tags: [Division]
      summary: List divisi (dengan paginasi)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/DivisionListPaged"
    post:
      tags: [Division]
      summary: Buat divisi baru
      description: "Butuh access_slug: `create_division`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/divisions/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathInt"
    get:
      tags: [Division]
      summary: Detail divisi
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/DivisionResponse" }
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Division]
      summary: Ubah divisi
      description: "Butuh access_slug: `edit_division`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Division]
      summary: Hapus divisi
      description: "Butuh access_slug: `delete_division`"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"

  # =========================================================
  # UNIT
  # =========================================================
  /api/units:
    get:
      tags: [Unit]
      summary: List satuan (dengan paginasi)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/UnitListPaged"
    post:
      tags: [Unit]
      summary: Buat satuan baru
      description: "Butuh access_slug: `create_unit`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/units/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathInt"
    get:
      tags: [Unit]
      summary: Detail satuan
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/UnitResponse" }
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Unit]
      summary: Ubah satuan
      description: "Butuh access_slug: `edit_unit`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Unit]
      summary: Hapus satuan
      description: "Butuh access_slug: `delete_unit`"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"

  # =========================================================
  # PPN
  # =========================================================
  /api/ppn:
    get:
      tags: [PPN]
      summary: List seluruh tarif PPN (histori)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/PpnResponse" }
    post:
      tags: [PPN]
      summary: Buat tarif PPN baru
      description: "Butuh access_slug: `edit_ppn`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PpnRequest" }
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/ppn/current:
    get:
      tags: [PPN]
      summary: Tarif PPN yang sedang berlaku
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PpnResponse" }
        "404":
          description: Tarif PPN belum diatur
  /api/ppn/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathInt"
    get:
      tags: [PPN]
      summary: Detail tarif PPN (untuk form edit)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PpnResponse" }
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [PPN]
      summary: Ubah tarif PPN
      description: "Butuh access_slug: `edit_ppn`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PpnRequest" }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"

  # =========================================================
  # ROLE
  # =========================================================
  /api/roles:
    get:
      tags: [Role]
      summary: List role (dengan paginasi, termasuk total_access)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/RoleResponse" }
    post:
      tags: [Role]
      summary: Buat role baru
      description: "Butuh access_slug: `create_role`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleRequest" }
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/roles/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathInt"
    get:
      tags: [Role]
      summary: Detail role beserta daftar access_id yang dimiliki
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/RoleDetailResponse" }
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Role]
      summary: Ubah role & daftar akses yang dimiliki
      description: "Butuh access_slug: `edit_role`"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleRequest" }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Role]
      summary: Hapus role
      description: "Butuh access_slug: `delete_role`"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"

  # =========================================================
  # ACCESS
  # =========================================================
...
  /api/access/module/{module}:
    get:
      tags: [Access]
      summary: List access berdasarkan modul (untuk form checklist per-modul)
      parameters:
        - name: module
          in: path
          required: true
          schema: { type: string }
          example: po
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/AccessResponse" }
...
  /api/admins/{id}/deactivate:
    parameters:
      - $ref: "#/components/parameters/IdPathString"
    post:
      tags: [Admin]
      summary: Nonaktifkan admin
      description: "Butuh access_slug: `deactivate_other_admin`"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    BadRequest:
      description: Permintaan tidak valid (validasi gagal)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponseBase" }
...
  schemas:
...
    ApiResponseBase:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
      example:
        success: true
        message: "Fetch success"
        data: {}
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
...
    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/TokenPairResponse"
        - type: object
          properties:
            admin: { $ref: "#/components/schemas/AdminResponse" }
...